fix(query): centralize SPARQL lexical admission
